+++ Modules/Collaborators/Http/Controllers/CollaboratorsApiController.php
@@ -3,27 +3,34 @@
 namespace Modules\Collaborators\Http\Controllers;
 
 use Exception;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
+use Modules\Core\Entities\User;
 use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\Log;
-use Modules\Collaborators\Http\Requests\UpdateCollaboratorRequest;
-use Modules\Collaborators\Transformers\CollaboratorsResource;
-use Modules\Core\Entities\User;
-use Modules\Collaborators\Http\Requests\StoreCollaboratorRequest;
+use Spatie\Activitylog\Models\Activity;
 use Vinkla\Hashids\Facades\Hashids;
+use Illuminate\Support\Facades\Gate;
+use Modules\Collaborators\Transformers\CollaboratorsResource;
+use Modules\Collaborators\Http\Requests\StoreCollaboratorRequest;
+use Modules\Collaborators\Http\Requests\UpdateCollaboratorRequest;
 
 class CollaboratorsApiController extends Controller
 {
     /**
-     * Display a listing of the resource.
-     * @return Response
+     * @return JsonResponse|AnonymousResourceCollection
      */
     public function index()
     {
         try {
             $userModel = new User();
             $user      = $userModel->where([['account_owner_id', auth()->user()->account_owner_id], ['id', '!=', auth()->user()->account_owner_id]]);
+
+            activity()->on($userModel)->tap(function(Activity $activity) {
+                $activity->log_name = 'visualization';
+            })->log('Visualizou tela todos colaboradores');
 
             return CollaboratorsResource::collection($user->orderBy('id', 'ASC')->paginate(5));
         } catch (Exception $e) {
@@ -40,7 +47,7 @@ class CollaboratorsApiController extends Controller
 
     /**
      * @param StoreCollaboratorRequest $request
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
      */
     public function store(StoreCollaboratorRequest $request)
     {
@@ -62,7 +69,9 @@ class CollaboratorsApiController extends Controller
             $data['score']                               = '0';
             $data['sms_zenvia_amount']                   = '0';
             $data['invites_amount']                      = 1;
-            $data['account_owner_id']                       = auth()->user()->account_owner_id;
+            $data['address_document_status']             = 3;
+            $data['personal_document_status']            = 3;
+            $data['account_owner_id']                    = auth()->user()->account_owner_id;
             $user                                        = $userModel->create($data);
             $user->assignRole($data['role']);
             if (!empty($user)) {
@@ -86,7 +95,7 @@ class CollaboratorsApiController extends Controller
 
     /**
      * @param $id
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
      */
     public function show($id)
     {
@@ -96,6 +105,19 @@ class CollaboratorsApiController extends Controller
 
             if ($userId) {
                 $user = $userModel->with('roles')->find($userId);
+
+                activity()->on($userModel)->tap(function(Activity $activity) use ($userId) {
+                    $activity->log_name   = 'visualization';
+                    $activity->subject_id = $userId;
+                })->log('Visualizou tela editar colaborador ' . $user->name);
+
+                if (Gate::denies('show', [$user])) {
+                    return response()->json([
+                                                'message' => 'Sem permissão',
+                                            ], Response::HTTP_FORBIDDEN
+                    );
+                }
+
                 if (!empty($user)) {
                     return new CollaboratorsResource($user);
                 } else {
@@ -123,7 +145,7 @@ class CollaboratorsApiController extends Controller
     /**
      * @param UpdateCollaboratorRequest $request
      * @param $id
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
      */
     public function update(UpdateCollaboratorRequest $request, $id)
     {
@@ -133,6 +155,14 @@ class CollaboratorsApiController extends Controller
             $userId    = current(Hashids::decode($id));
             if ($userId) {
                 $user = $userModel->find($userId);
+
+                if (Gate::denies('update', [$user])) {
+                    return response()->json([
+                                                'message' => 'Sem permissão',
+                                            ], Response::HTTP_FORBIDDEN
+                    );
+                }
+
                 if (!empty($data['password'])) {
                     $data['password'] = bcrypt($data['password']);
                 }
@@ -166,14 +196,22 @@ class CollaboratorsApiController extends Controller
 
     /**
      * @param $id
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
      */
     public function destroy($id)
     {
         $userModel = new User();
         $userId    = current(Hashids::decode($id));
         if ($userId) {
-            $user        = $userModel->find($userId);
+            $user = $userModel->find($userId);
+
+            if (Gate::denies('destroy', [$user])) {
+                return response()->json([
+                                            'message' => 'Sem permissão',
+                                        ], Response::HTTP_FORBIDDEN
+                );
+            }
+
             $userDeleted = $user->delete();
             if ($userDeleted) {
                 return response()->json([
